+++ lib/braintree/transaction_gateway.rb
@@ -303,6 +303,12 @@ module Braintree
             },
           ]
         },
+        {
+          :us_bank_account => [
+            :ach_mandate_text,
+            :ach_mandate_accepted_at,
+          ]
+        },
       ]
     end