+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Add `remaining_file_evidence_storage` to Dispute
 - Add `transfer_type` to Transaction
 - Add `transaction_retried` webhook
+- Add Bank Account Instant Verification Gateway with JWT creation support
+- Add ACH mandate support to transactions with `ach_mandate_text` and `ach_mandate_accepted_at` fields in `us_bank_account`
+- Add `InstantVerification` method to `UsBankAccountVerification::VerificationMethod`
 
 ## 4.29.0